refactor: rename configuration files and update paths
Refactor configuration files to have consistent naming conventions.
This is necessary to avoid problems with existing plugins configuration names.
