+++ scripts/oxlint-compatgen.ts
@@ -2,7 +2,7 @@ import fs from 'node:fs/promises';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import * as categoriesRules from 'eslint-plugin-oxlint/rules-by-category';
-import { config } from '../src/eslint.config';
+import { config } from '../src/eslint/config';
 import { BASE_IGNORES } from '../src/eslint/configs/base';
 
 // Define the plugins scopes on oxlint with their possible scopes for eslint
@@ -122,7 +122,7 @@ for (const c of eslintConfigs) {
 // Write `oxlint.config.jsonc` file
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 await fs.writeFile(
-  `${__dirname}/../src/oxlint.config.jsonc`,
+  `${__dirname}/../src/config.oxlint.jsonc`,
   JSON.stringify({
     $schema:
       'https://raw.githubusercontent.com/oxc-project/oxc/main/npm/oxlint/configuration_schema.json',
